refactor(default layout): split server code from client code
