+++ frontend/layouts/default/Newsletter.tsx
