--- frontend/layouts/default/index.ts
@@ -1,2 +1,0 @@
-export * from './DefaultLayout';
-export * from './getLayoutServerSideProps';
