+++ frontend/templates/product/server.tsx
@@ -6,7 +6,7 @@ import {
 import { ifixitOriginFromHost } from '@helpers/path-helpers';
 import { invariant } from '@ifixit/helpers';
 import { urlFromContext } from '@ifixit/helpers/nextjs';
-import { getLayoutServerSideProps } from '@layouts/default';
+import { getLayoutServerSideProps } from '@layouts/default/server';
 import { findProduct } from '@models/product';
 import { GetServerSideProps } from 'next';
 import { ProductTemplateProps } from './hooks/useProductTemplateProps';
